Implementação de consulta de livros by filter paginada com Pageable com TDD. Testes de Integração(camada repository). Testes unitários para a camada Service. Testes unitários para a camada controller.

+++ src/main/java/com/carledwinti/library/api/service/impl/BookServiceImpl.java
@@ -5,6 +5,8 @@ import com.carledwinti.library.api.exception.BusinessException;
 import com.carledwinti.library.api.model.Book;
 import com.carledwinti.library.api.repository.BookRepository;
 import com.carledwinti.library.api.service.BookService;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.Optional;
@@ -28,16 +30,30 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public Optional<Book> getByid(Long id) {
+        if(id == null){
+            throw new IllegalArgumentException(ConstantsError.MSG_ERROR_ID_CANT_BE_NULL);
+        }
         return bookRepository.findById(id);
     }
 
     @Override
     public void delete(Book book) {
+        if(book == null || book.getId() == null){
+            throw new IllegalArgumentException(ConstantsError.MSG_ERROR_BOOK_AND_ID_CANT_BE_NULL);
+        }
         bookRepository.delete(book);
     }
 
     @Override
     public Book update(Book book) {
+        if(book == null || book.getId() == null){
+            throw new IllegalArgumentException(ConstantsError.MSG_ERROR_BOOK_AND_ID_CANT_BE_NULL);
+        }
         return bookRepository.save(book);
+    }
+
+    @Override
+    public Page<Book> find(Book bookFilter, Pageable pageRequest) {
+        return null;
     }
 }
